requirements.txt
fastapi
pydantic
mayat
GitPython
pika
sqlalchemy